development of README, reorganised layout as a package. working on readme, made a file of all used sql statements,
