+++ flaskr/templates/layout_login_registration.html
