+++ flaskr/templates/list.html
@@ -56,6 +56,7 @@
       </div>
     </div>
   </Fieldset>
+  <input type="hidden" name="csrf_token" value="{{ csrf_token() }}">
 </form>
 
 {% endblock %}
