+++ flaskr/templates/my_account.html
@@ -115,6 +115,7 @@
     <div class="form-group">
       {{ form.submit(class="btn btn-success") }}
     </div>
+    <input type="hidden" name="csrf_token" value="{{ csrf_token() }}">
   </form>
 
   <small class="d-block text-right mt-3 mb-3">
