+++ flaskr/templates/my_lists.html
@@ -66,6 +66,7 @@
       </div>
     </div>
   </Fieldset>
+  <input type="hidden" name="csrf_token" value="{{ csrf_token() }}">
 </form>
