+++ flaskr/templates/register.html
@@ -75,6 +75,7 @@
     <div class="form-group">
       {{ form.submit(class="btn btn-success mt-1 mb-1") }}
     </div>
+    <input type="hidden" name="csrf_token" value="{{ csrf_token() }}">
   </form>
 </div>
